Minor changes, changed how QSnapper functions.
QSnapper now has the option to allow "slightly different" images, so
things like a blinking text cursor won't create a new image. The
threshold for new images is that the pixels changed must be more than 1%
of the display's pixels: (width * height)/100.

Also added the ability to view "difference images", where the
differences between two images are drawn on top of a light blue, so you
can see exactly what changed, if it seems like two images are identical.